# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: sqi-server API
  description: |
    REST API for the sqi distributed task management server.

    All error responses use the RFC 7807 problem-details format
    (`application/problem+json`).  Clients should branch on `Content-Type`
    rather than status code alone to detect error bodies.
  version: "0.1.0"
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /api/v1
    description: Current server
...
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource UUID
...
  schemas:
...
    ProblemDetail:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          example: "about:blank"
        title:
          type: string
          example: Not Found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: job not found
        instance:
          type: string
          description: Request-ID that identifies this specific occurrence.
          example: a1b2c3d4e5f60708

    # ── Jobs ─────────────────────────────────────────────────────────────────
...
    Farm:
      type: object
      required: [id, name, max_concurrent_tasks, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        max_concurrent_tasks:
          type: integer
          minimum: 0
          description: 0 means unlimited.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateFarmRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        max_concurrent_tasks:
          type: integer
          minimum: 0
          default: 0
    UpdateFarmRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        max_concurrent_tasks:
          type: integer
          minimum: 0
          default: 0

    # ── Queues ────────────────────────────────────────────────────────────────
...
  responses:
    BadRequest:
      description: Bad request — missing or invalid parameters.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Conflict:
      description: Conflict — unique constraint violated or invalid state transition.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    InternalServerError:
      description: Internal server error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /farms:
    post:
      operationId: createFarm
      summary: Create a farm
      tags: [farms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFarmRequest"
      responses:
        "201":
          description: Farm created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      operationId: listFarms
      summary: List all farms
      tags: [farms]
      responses:
        "200":
          description: Array of farms.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farm"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /farms/{id}:
    parameters:
      - $ref: "#/components/parameters/id"

    get:
      operationId: getFarm
      summary: Get a farm
      tags: [farms]
      responses:
        "200":
          description: Farm.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      operationId: updateFarm
      summary: Update a farm
      tags: [farms]
      description: Full replacement of mutable fields.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFarmRequest"
      responses:
        "200":
          description: Updated farm.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      operationId: deleteFarm
      summary: Delete a farm
      tags: [farms]
      responses:
        "204":
          description: Farm deleted.
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ── Queues ─────────────────────────────────────────────────────────────────
...